+++ D2D/Forest/Spiker.cpp
@@ -32,7 +32,7 @@ void Spiker::Collision(shared_ptr<Player> player)
 		{
 			player->SetJumpSpeed(400.0f);
 			player->SetG(0.0f);
-			player->SetSuperMeterCard(player->GetSuperMeterCard() + 0.2 * player->GetMaxSuperMeterCard());	// 20 퍼센트 추가
+			player->SetSuperMeterCard((float)(player->GetSuperMeterCard() + 0.2 * player->GetMaxSuperMeterCard()));	// 20 퍼센트 추가
 			bActivation = false;
 			parryTime = 0.0f;
 		}
